PhotoVault/python_sidecar/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
ultralytics>=8.1.0
insightface>=0.7.3
onnxruntime-gpu>=1.17.0
open-clip-torch>=2.24.0
paddleocr>=2.7.0
paddlepaddle-gpu>=2.6.0
Pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.9
